workspace

+++ src/main/java/Discount.kt
@@ -2,6 +2,6 @@ class Discount {
 
     var name     = String;
     var type     = String;
-    var value    = Int;
+    var value    = Double;
 
 }
